Provide access to some physical memory via PMPEP1.
Applications on Linux can allocate memory from their resource manager,
which takes it from the pool that has been assigned to the tile. Most
of the time, this is accessible through PMPEP1, which is why we
configure RISC-V's PMP to include some range accessible via PMPEP1.

That's not nice, but enough for our purposes right now and only a
debugging feature anyway.