Added optional line buffering to UART.
This prevents that we see mixtures of output from, e.g., different Linux
instances on a character level and also improves performance. It can be
disabled, because it does not work well when using the system
interactively as there is no explicit flush operation, so that we only
flush on newline.